+++ src/torchjd/scalarization/__init__.py
@@ -24,6 +24,7 @@ from ._geometric_mean import GeometricMean
 from ._mean import Mean
 from ._random import Random
 from ._scalarizer_base import Scalarizer
+from ._stch import STCH
 from ._sum import Sum
 
-__all__ = ["Constant", "GeometricMean", "Mean", "Random", "Scalarizer", "Sum"]
+__all__ = ["Constant", "GeometricMean", "Mean", "Random", "Scalarizer", "STCH", "Sum"]